Move color classes to particular directory
